Don't send decrypted data over the wire
Before this patch, everything the client received on stdin was sent to the
server for decryption. This has bad security properties. Namely, it means
that the Petera server (or anything that compromises TLS) can see the
decrypted data.

Instead, during encryption we generate a random key. This random key is
encrypted to the Petera server. Then the random key is used to encrypt the
data locally. Any compromise of the Petera server or its TLS will only leak
this random key, not the input data. An attacker would then also have to
compromise the client system to access the local data.

Two other benefits arise from this more secure implementation:
1. No more size limits on encryption/decryption.
2. Network packet size doesn't vary with input size.

This patch changes the Petera IO format and network protocol without
backwards compatibility.