Huge refactor to allow for parameter-less decryption
The basic gist of this huge patch is that we now embed the anchors
into the PETERA_HEADER. This means that when performing decryption,
no parameters are required.

This patch also changes the input commands such that anchors are
now specified with the -a flag. Mutliple anchors can be given, but
directories are no longer supported. If no anchors are specified,
the system default trust anchors are used.